+++ Frends.AzureBlobStorage.ListBlobsInContainer/Frends.AzureBlobStorage.ListBlobsInContainer.Tests/ConnectionStringUnitTests.cs
@@ -1,14 +1,10 @@
-﻿using Azure.Storage.Blobs;
-using Azure.Storage.Blobs.Models;
-using Frends.AzureBlobStorage.ListBlobsInContainer.Definitions;
+﻿using Frends.AzureBlobStorage.ListBlobsInContainer.Definitions;
+using Frends.AzureBlobStorage.ListBlobsInContainer.Tests.lib;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
-using System.IO;
 using System.Linq;
-using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 
 namespace Frends.AzureBlobStorage.ListBlobsInContainer.Tests;
@@ -22,13 +18,13 @@ public class ConnectionStringUnitTests
     [TestInitialize]
     public async Task Init()
     {
-        await CreateContainerAndTestFiles(false);
+        await Helper.CreateContainerAndTestFiles(false, _connstring, _containerName);
     }
 
     [TestCleanup]
     public async Task CleanUp()
     {
-        await CreateContainerAndTestFiles(true);
+        await Helper.CreateContainerAndTestFiles(true, _connstring, _containerName);
     }
 
     [TestMethod]
@@ -43,7 +39,7 @@ public class ConnectionStringUnitTests
             ContainerName = _containerName
         };
 
-        foreach(var structure in listing) 
+        foreach (var structure in listing)
         {
             var options = new Options
             {
@@ -53,7 +49,7 @@ public class ConnectionStringUnitTests
 
             var result = await AzureBlobStorage.ListBlobsInContainer(source, options, default);
 
-            if(structure is ListingStructure.Flat)
+            if (structure is ListingStructure.Flat)
             {
                 Assert.IsTrue(result.BlobList.Any(x => x.Name == "Temp/SubFolderFile"));
                 Assert.IsTrue(result.BlobList.Any(x => x.Name == "Temp/SubFolderFile2"));
@@ -74,7 +70,7 @@ public class ConnectionStringUnitTests
             Assert.IsTrue(result.BlobList.Any(x => x.LastModified != null));
         }
     }
-    
+
     [TestMethod]
     public async Task ListBlob_ConnectionString_Prefix()
     {
@@ -129,54 +125,7 @@ public class ConnectionStringUnitTests
             ListingStructure = ListingStructure.Hierarchical
         };
 
-        var ex = await Assert.ThrowsExceptionAsync<Exception>(async () => await AzureBlobStorage.ListBlobsInContainer(source, options, default));
+        var ex = await Assert.ThrowsExceptionAsync<ArgumentException>(async () => await AzureBlobStorage.ListBlobsInContainer(source, options, default));
         Assert.AreEqual("Connection string required.", ex.InnerException.Message);
-    }
-
-    private async Task CreateContainerAndTestFiles(bool delete)
-    {
-        var blobServiceClient = new BlobServiceClient(_connstring);
-        var container = blobServiceClient.GetBlobContainerClient(_containerName);
-        if (delete)
-            await container.DeleteIfExistsAsync();
-        else
-        {
-            await container.CreateIfNotExistsAsync(PublicAccessType.None, null, null, new CancellationToken());
-
-            byte[] bytes;
-
-            try
-            {
-                var files = new List<string>()
-                {
-                    "TestFile.txt", "TestFile2.txt", "Temp/SubFolderFile", "Temp/SubFolderFile2"
-                };
-
-
-                foreach(var file in files )
-                {
-                    Directory.CreateDirectory(Path.Combine(Directory.GetCurrentDirectory(), "Temp"));
-
-                    var tempFile = Directory.GetCurrentDirectory() + "/" + file;
-                    using (StreamWriter sw = File.CreateText(tempFile))
-                        sw.WriteLine($"This is {file}");
-                    
-                    using (var reader = new StreamReader(tempFile)) 
-                        bytes = Encoding.UTF32.GetBytes(reader.ReadToEnd());
-                    
-                    await container.UploadBlobAsync(file, new MemoryStream(bytes));
-                    
-                    if(File.Exists(tempFile))
-                        File.Delete(tempFile);
-
-                    if (Directory.Exists(Path.Combine(Directory.GetCurrentDirectory(), "Temp")))
-                        Directory.Delete(Path.Combine(Directory.GetCurrentDirectory(), "Temp"), true);
-                }
-            }
-            catch (Exception ex)
-            {
-                throw new Exception(ex.Message);
-            }
-        }
     }
 }