+++ Frends.AzureBlobStorage.ListBlobsInContainer/Frends.AzureBlobStorage.ListBlobsInContainer.Tests/SASUnitTests.cs
@@ -1,23 +1,20 @@
 ﻿using Azure.Storage;
-using Azure.Storage.Blobs;
-using Azure.Storage.Blobs.Models;
 using Azure.Storage.Sas;
 using Frends.AzureBlobStorage.ListBlobsInContainer.Definitions;
+using Frends.AzureBlobStorage.ListBlobsInContainer.Tests.lib;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
-using System.IO;
 using System.Linq;
-using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
+
 
 namespace Frends.AzureBlobStorage.ListBlobsInContainer.Tests;
 
 [TestClass]
 public class SASUnitTests
-{ 
+{
     private readonly string _accessKey = Environment.GetEnvironmentVariable("HiQ_AzureBlobStorage_testsorage01AccessKey");
     private readonly string _connstring = Environment.GetEnvironmentVariable("HiQ_AzureBlobStorage_ConnString");
     private readonly string _containerName = $"test-container{DateTime.Now.ToString("mmssffffff", CultureInfo.InvariantCulture)}";
@@ -26,13 +23,13 @@ public class SASUnitTests
     [TestInitialize]
     public async Task Init()
     {
-        await CreateContainerAndTestFiles(false);
+        await Helper.CreateContainerAndTestFiles(false, _connstring, _containerName);
     }
 
     [TestCleanup]
     public async Task CleanUp()
     {
-        await CreateContainerAndTestFiles(true);
+        await Helper.CreateContainerAndTestFiles(true, _connstring, _containerName);
     }
 
     [TestMethod]
@@ -44,7 +41,6 @@ public class SASUnitTests
             URI = _uri,
             SASToken = "",
             ContainerName = _containerName,
-
         };
 
         var options = new Options
@@ -53,7 +49,7 @@ public class SASUnitTests
             ListingStructure = ListingStructure.Flat,
         };
 
-        var ex = await Assert.ThrowsExceptionAsync<Exception>(async () => await AzureBlobStorage.ListBlobsInContainer(source, options, default));
+        var ex = await Assert.ThrowsExceptionAsync<ArgumentException>(async () => await AzureBlobStorage.ListBlobsInContainer(source, options, default));
         Assert.AreEqual("SAS Token and URI required.", ex.InnerException.Message);
     }
 
@@ -141,53 +137,6 @@ public class SASUnitTests
         }
     }
 
-    private async Task CreateContainerAndTestFiles(bool delete)
-    {
-        var blobServiceClient = new BlobServiceClient(_connstring);
-        var container = blobServiceClient.GetBlobContainerClient(_containerName);
-        if (delete)
-            await container.DeleteIfExistsAsync();
-        else
-        {
-            await container.CreateIfNotExistsAsync(PublicAccessType.None, null, null, new CancellationToken());
-
-            byte[] bytes;
-
-            try
-            {
-                var files = new List<string>()
-                {
-                    "TestFile.txt", "TestFile2.txt", "Temp/SubFolderFile", "Temp/SubFolderFile2"
-                };
-
-
-                foreach(var file in files )
-                {
-                    Directory.CreateDirectory(Path.Combine(Directory.GetCurrentDirectory(), "Temp"));
-
-                    var tempFile = Directory.GetCurrentDirectory() + "/" + file;
-                    using (StreamWriter sw = File.CreateText(tempFile))
-                        sw.WriteLine($"This is {file}");
-                    
-                    using (var reader = new StreamReader(tempFile)) 
-                        bytes = Encoding.UTF32.GetBytes(reader.ReadToEnd());
-                    
-                    await container.UploadBlobAsync(file, new MemoryStream(bytes));
-                    
-                    if(File.Exists(tempFile))
-                        File.Delete(tempFile);
-
-                    if (Directory.Exists(Path.Combine(Directory.GetCurrentDirectory(), "Temp")))
-                        Directory.Delete(Path.Combine(Directory.GetCurrentDirectory(), "Temp"), true);
-                }
-            }
-            catch (Exception ex)
-            {
-                throw new Exception(ex.Message);
-            }
-        }
-    }
-
     private string GenerateSASToken()
     {
         BlobSasBuilder blobSasBuilder = new()
@@ -195,7 +144,7 @@ public class SASUnitTests
             BlobContainerName = _containerName,
             ExpiresOn = DateTime.UtcNow.AddMinutes(5)
         };
-        //blobSasBuilder.SetPermissions(BlobAccountSasPermissions.List);
+
         blobSasBuilder.SetPermissions(BlobContainerSasPermissions.List);
         var sasToken = blobSasBuilder.ToSasQueryParameters(new StorageSharedKeyCredential("testsorage01", _accessKey)).ToString();
         return sasToken;